Handeling sparse matrix in the old way, moving all int to size_t and restructuring cuda

+++ bioinf_learn/computation/inverseIndexCuda.h
@@ -17,54 +17,55 @@
 class InverseIndexCuda {
 
   private: 
-    int* mDev_FeatureList;
-    int* mDev_SizeOfInstanceList;
-    int* mDev_ComputedSignaturesPerInstance;
-    float* mDev_ValuesList;
-    int* mDev_JumpLength;
-    float* mDev_DotProduct;
-    int mNumberOfHashFunctions;
-    int mShingleSize;
-    int mShingle;
-    int mHashAlgorithm;
-    int mBlockSize;
+    size_t* mDev_FeatureList;
+    size_t* mDev_SizeOfInstanceList;
+    size_t* mDev_ComputedSignaturesPerInstance;
+    size_t* mDev_ValuesList;
+    size_t* mDev_JumpLength;
+    size_t* mDev_DotProduct;
+    
+    size_t mNumberOfHashFunctions;
+    size_t mShingleSize;
+    size_t mShingle;
+    size_t mHashAlgorithm;
+    size_t mBlockSize;
     bool mDataOnGpu;
   public:
-  	InverseIndexCuda(int pNumberOfHashFunctions, int pShingle, 
-                        int pShingleSize, int pBlockSize, int pHashAlgorithm);
+  	InverseIndexCuda(size_t pNumberOfHashFunctions, size_t pShingle, 
+                        size_t pShingleSize, size_t pBlockSize, size_t pHashAlgorithm);
     ~InverseIndexCuda();
-    void copyDataToGpu(SparseMatrixFloat* pRawData, int** pDevFeatureList,
-                                      float** pDevValueList, int** pSizeList, int** pJumpList);
+    void copyDataToGpu(SparseMatrixFloat* pRawData, size_t** pDevFeatureList,
+                                      size_t** pDevValueList, size_t** pSizeList, size_t** pJumpList);
   	void computeSignaturesFittingOnGpu(SparseMatrixFloat* pRawData, 
-                                    int pStartIndex, int pEndIndex, 
-                                    int pNumberOfInstances, int pNumberOfBlocks, 
-                                    int pNumberOfThreads, int pShingleFactor, 
-                                    int pBlockSizeShingle,
-                                    vvsize_t_p* pSignatures, int pRangeK);
+                                        size_t pStartIndex, size_t pEndIndex, 
+                                        size_t pNumberOfInstances, size_t pNumberOfBlocks, 
+                                        size_t pNumberOfThreads, size_t pShingleFactor, 
+                                        size_t pBlockSizeShingle,
+                                        vvsize_t_p* pSignatures, size_t pRangeK);
     // void copyFittingDataToGpu(SparseMatrixFloat* pRawData, size_t pStartIndex);
     
    void computeSignaturesQueryOnGpu(SparseMatrixFloat* pRawData, 
-                                                int pStartIndex, int pEndIndex, 
-                                                int pNumberOfInstances, int pNumberOfBlocks, 
-                                                int pNumberOfThreads, int pShingleFactor, 
-                                                int pBlockSizeShingle,
-                                                vvsize_t_p* pSignatures, int pRangeK);
-   int** get_mDev_FeatureList() {
+                                                size_t pStartIndex, size_t pEndIndex, 
+                                                size_t pNumberOfInstances, size_t pNumberOfBlocks, 
+                                                size_t pNumberOfThreads, size_t pShingleFactor, 
+                                                size_t pBlockSizeShingle,
+                                                vvsize_t_p* pSignatures, size_t pRangeK);
+   size_t** get_mDev_FeatureList() {
        return &mDev_FeatureList;
    };
-   int** get_mDev_SizeOfInstanceList() {
+   size_t** get_mDev_SizeOfInstanceList() {
        return &mDev_SizeOfInstanceList;
    };
-   int** get_mDev_ComputedSignaturesPerInstance() {
+   size_t** get_mDev_ComputedSignaturesPerInstance() {
        return &mDev_ComputedSignaturesPerInstance;
    };
-   float** get_mDev_ValuesList() {
+   size_t** get_mDev_ValuesList() {
        return &mDev_ValuesList;
    };
-   int** get_mDev_JumpLength() {
+   size_t** get_mDev_JumpLength() {
        return &mDev_JumpLength;
    };
-   float** get_mDev_DotProduct() {
+   size_t** get_mDev_DotProduct() {
        return &mDev_DotProduct;
    };
 };
